fastmcp-agents-library/agents/fastmcp-agents-library-agent-filesystem-operations/pyproject.toml
[project]
name = "fastmcp-agents-library-agent-filesystem-operations"
version = "0.5.6"
description = "An agent for interacting with the filesystem"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp-agents-library-mcp",
    "fastmcp-ai-agent-bridge-pydantic-ai>=0.1.0",
]

[tool.uv.sources]
fastmcp-agents-library-mcp = { workspace = true }

[build-system]
requires = ["uv_build>=0.8.2,<0.9.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-name = "fastmcp_agents.library.agent.filesystem_operations"

[dependency-groups]
dev = [
    "ruff>=0.12.4",
    "pytest>=8.3.4",
    "pytest-asyncio>=0.24.0",
]

[project.scripts]
filesystem-agent = "fastmcp_agents.library.agent.filesystem_operations.agent:run"

[tool.ruff]
extend="../../../pyproject.toml"

[tool.pyright]
extends = "../../../pyproject.toml"
